+++ NEWS.md
@@ -5,6 +5,7 @@
 
 - Add security posture.
 - Fix URL escaping of non-ASCII bytes.
+- Fix environment restoration during re-entrant template rendering.
 
 
 ## Release 1.0.0 (2024-04-06)